% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr.R
\name{assemble_data}
\alias{assemble_data}
\title{Assemble Data}
\usage{
assemble_data()
}
\value{
A data.frame that contains the interest rates, points and fees, and
(for adjustable-rate loans) the margin for different term loans.
}
\description{
This function compiles all of the information necessary to generate Average
Prime Offer Rates (APORs).  This information includes values from the Freddie
Mac Primary Mortgage Market Survey (PMMS) and U.S. Treasury yields. In addition
to these data sources, the function also imputes values for intermediate terms
as specified in the procedure for generating APORs.
}
