% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr.R
\name{calculate_apors}
\alias{calculate_apors}
\title{Calculate APORs}
\usage{
calculate_apors(dt, type)
}
\arguments{
\item{dt}{A data.frame of the type produced by assemble_date()}

\item{type}{A character vector indicating if APORs are to be calculated for
fixed- or adjustable-rate loans (e.g., 'fixed', 'arm')}
}
\description{
This function calculates the Average Prime Offer Rates (APORs) for the dates
for which data have been supplied.
}
