% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr.R
\name{calculate_apr_fixed}
\alias{calculate_apr_fixed}
\title{Fixed APR Calculation}
\usage{
calculate_apr_fixed(r, p, t = 360L)
}
\arguments{
\item{r}{Annual interest rate as a percentage}

\item{p}{Points and Fees expressed as percentage of loan amount}

\item{t}{Term of loan in months}
}
\description{
This function computes the APR for a fixed-rate loan.
}
