% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr.R
\name{calculate_monthly_payment}
\alias{calculate_monthly_payment}
\title{Calculate Monthly Payment}
\usage{
calculate_monthly_payment(r, t, l = 1L)
}
\arguments{
\item{r}{Annual interest rate as a percentage}

\item{t}{Term of the loan in months}

\item{l}{Loan amount (defaults to $1)}
}
\value{
The monthly payment required to fully payoff the loan by the end of the term.
}
\description{
This function calculates the monthly payment required to fully pay off a loan
with a fixed term and a constant interest rate.
}
