% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmms.R
\name{retrieve_pmms_data}
\alias{retrieve_pmms_data}
\title{Retrieve PMMS Data}
\usage{
retrieve_pmms_data()
}
\description{
This function will retrieve the necessary survey data to calculate Average Percentage
Offer Rates (APORs).
}
