% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasury.R
\name{retrieve_tbills}
\alias{retrieve_tbills}
\title{Retrieve Treasury Data}
\usage{
retrieve_tbills()
}
\description{
This function will pull down the necessary data on Treasury yields and return
a data.frame with the correct average T-bill rates for each PMMS date.
}
