% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr.R
\name{round_higher}
\alias{round_higher}
\title{Round Midpoint Values Up}
\usage{
round_higher(x, y, term, digits)
}
\arguments{
\item{x}{Value based on 1-year observations (Treasury or PMMS)}

\item{y}{Value based on 5-year observations}

\item{term}{Character value indicating if the rounding is being done for the
2-year or 3-year products.}

\item{digits}{The number of digits to include in the value returned}
}
\value{
The weighted average of x and y
}
\description{
R's default rounding behavior is to round values of .5 to the nearest even
integer. This creates inconsistencies with SAS, which rounds up in all cases.
To match the published values of APOR, this function will round midpoint values
up in all cases.
}
